Separating food goals from foods
