+++ src/Food/Foods/Foods.php
@@ -28,18 +28,6 @@ class Foods
     }
 
     /**
-     * Returns a user's current daily calorie consumption goal and/or food Plan in the format requested.
-     */
-    public function getGoals()
-    {
-        return $this->fitbit->get(implode('/', [
-            'foods',
-            'log',
-            'goal',
-          ]) . '.json');
-    }
-
-    /**
      * Returns a summary and list of a user's food log entries for a given day in the format requested.
      *
      * @param Carbon $date
